refactor: extract common modal component for gallery and video
